Make feed id autoincrement
This is necessary to avoid a condition where the most recently added
feed is deleted, and a new one subsequently added, after the latest id
has been fetched from the database for updating but before the new
entries are inserted. In this case, without autoincrement, entries from
the deleted feed would be associated with the newly added feed.

This fix does not include migration from the old DB schema without
the change since this project is not stable yet, but I intend to start
including such migrations in perhaps the very next release that alters
the database schema.